+++ client/web/src/search/results/sidebar/SearchReference.tsx
@@ -4,7 +4,6 @@ import { escapeRegExp } from 'lodash'
 import ChevronDownIcon from 'mdi-react/ChevronDownIcon'
 import ChevronLeftIcon from 'mdi-react/ChevronLeftIcon'
 import ExternalLinkIcon from 'mdi-react/ExternalLinkIcon'
-import { Range, Selection } from 'monaco-editor'
 import React, { ReactElement, useCallback, useMemo, useState } from 'react'
 import { Collapse } from 'reactstrap'
 
@@ -13,8 +12,6 @@ import { Markdown } from '@sourcegraph/shared/src/components/Markdown'
 import { SearchPatternType } from '@sourcegraph/shared/src/graphql-operations'
 import { FILTERS, FilterType, isNegatableFilter } from '@sourcegraph/shared/src/search/query/filters'
 import { scanSearchQuery } from '@sourcegraph/shared/src/search/query/scanner'
-import { appendFilter, updateFilter } from '@sourcegraph/shared/src/search/query/transformer'
-import { findFilter, FilterKind } from '@sourcegraph/shared/src/search/query/validate'
 import { VersionContextProps } from '@sourcegraph/shared/src/search/util'
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
 import { renderMarkdown } from '@sourcegraph/shared/src/util/markdown'
@@ -22,15 +19,15 @@ import { useLocalStorage } from '@sourcegraph/shared/src/util/useLocalStorage'
 
 import { CaseSensitivityProps, PatternTypeProps, SearchContextProps } from '../..'
 import { QueryChangeSource, QueryState } from '../../helpers'
+import { createQueryExampleFromString, updateQueryWithFilterAndExample, QueryExample } from '../../helpers/queryExample'
 
 import styles from './SearchReference.module.scss'
 import sidebarStyles from './SearchSidebarSection.module.scss'
 
 const SEARCH_REFERENCE_TAB_KEY = 'SearchProduct.SearchReference.Tab'
 
-export interface FilterInfo {
-    type: FilterType
-    placeholder: Placeholder
+type FilterInfo = QueryExample & {
+    field: FilterType
     description: string
     /**
      * Force showing or not showing suggestions for this fileter.
@@ -45,9 +42,8 @@ export interface FilterInfo {
     examples?: string[]
 }
 
-interface OperatorInfo {
+type OperatorInfo = QueryExample & {
     operator: string
-    placeholder: Placeholder
     description: string
     alias?: string
     examples?: string[]
@@ -59,7 +55,7 @@ type SearchReferenceInfo = FilterInfo | OperatorInfo
  * Adds additional search reference information from the existing filters list.
  */
 function augmentFilterInfo(searchReference: FilterInfo): void {
-    const filter = FILTERS[searchReference.type]
+    const filter = FILTERS[searchReference.field]
     if (filter?.alias) {
         searchReference.alias = filter.alias
     }
@@ -67,23 +63,23 @@ function augmentFilterInfo(searchReference: FilterInfo): void {
 
 const filterInfos: FilterInfo[] = [
     {
-        type: FilterType.after,
-        placeholder: parsePlaceholder('"{last week}"'),
+        ...createQueryExampleFromString('"{last week}"'),
+        field: FilterType.after,
         description:
             'Only include results from diffs or commits which have a commit date after the specified time frame. To use this filter, the search query must contain `type:diff` or `type:commit`.',
         commonRank: 100,
         examples: ['after:"6 weeks ago"', 'after:"november 1 2019"'],
     },
     {
-        type: FilterType.archived,
-        placeholder: parsePlaceholder('{yes/only}'),
+        ...createQueryExampleFromString('{yes/only}'),
+        field: FilterType.archived,
         description:
             'The "yes" option includes archived repositories. The "only" option filters results to only archived repositories. Results in archived repositories are excluded by default.',
         examples: ['repo:sourcegraph/ archived:only'],
     },
     {
-        type: FilterType.author,
-        placeholder: parsePlaceholder('{name}'),
+        ...createQueryExampleFromString('{name}'),
+        field: FilterType.author,
         description: `Only include results from diffs or commits authored by the user. Regexps are supported. Note that they match the whole author string of the form \`Full Name <user@example.com>\`, so to include only authors from a specific domain, use \`author:example.com>$\`.
 
 You can also search by \`committer:git-email\`. *Note: there is a committer only when they are a different user than the author.*
@@ -92,74 +88,74 @@ To use this filter, the search query must contain \`type:diff\` or \`type:commit
         examples: ['type:diff author:nick'],
     },
     {
-        type: FilterType.before,
-        placeholder: parsePlaceholder('"{last thursday}"'),
+        ...createQueryExampleFromString('"{last thursday}"'),
+        field: FilterType.before,
         description:
             'Only include results from diffs or commits which have a commit date before the specified time frame. To use this filter, the search query must contain `type:diff` or `type:commit`.',
         commonRank: 100,
         examples: ['before:"last thursday"', 'before:"november 1 2019"'],
     },
     {
-        type: FilterType.case,
-        placeholder: parsePlaceholder('{yes}'),
+        ...createQueryExampleFromString('{yes}'),
+        field: FilterType.case,
         description: 'Perform a case sensitive query. Without this, everything is matched case insensitively.',
         examples: ['OPEN_FILE case:yes'],
     },
     {
-        type: FilterType.content,
-        placeholder: parsePlaceholder('"{pattern}"'),
+        ...createQueryExampleFromString('"{pattern}"'),
+        field: FilterType.content,
         description:
             'Set the search pattern with a dedicated parameter. Useful when searching literally for a string that may conflict with the search pattern syntax. In between the quotes, the `\\` character will need to be escaped (`\\\\` to evaluate for `\\`).',
         commonRank: 70,
         examples: ['repo:sourcegraph content:"repo:sourcegraph"', 'file:Dockerfile alpine -content:alpine:latest'],
     },
     {
-        type: FilterType.count,
-        placeholder: parsePlaceholder('{N/all}'),
+        ...createQueryExampleFromString('{N/all}'),
+        field: FilterType.count,
         description:
             'Retrieve *N* results. By default, Sourcegraph stops searching early and returns if it finds a full page of results. This is desirable for most interactive searches. To wait for all results, use **count:all**.',
         commonRank: 60,
         examples: ['count:1000 function', 'count:all err'],
     },
     {
-        type: FilterType.file,
-        placeholder: parsePlaceholder('{regexp-pattern}'),
+        ...createQueryExampleFromString('{regexp-pattern}'),
+        field: FilterType.file,
         commonRank: 30,
         description: 'Only include results in files whose full path matches the regexp.',
         examples: ['file:.js$ httptest', 'file:internal/ httptest', 'file:.js$ -file:test http'],
     },
     {
-        type: FilterType.file,
-        placeholder: parsePlaceholder('contains.content({regexp-pattern})'),
+        ...createQueryExampleFromString('contains.content({regexp-pattern})'),
+        field: FilterType.file,
         description: 'Search only inside files that contain content matching the provided regexp pattern.',
         examples: ['file:contains.content(github.com/sourcegraph/sourcegraph)'],
     },
     {
-        type: FilterType.fork,
-        placeholder: parsePlaceholder('{yes/only}'),
+        ...createQueryExampleFromString('{yes/only}'),
+        field: FilterType.fork,
         description:
             'Include results from repository forks or filter results to only repository forks. Results in repository forks are exluded by default.',
         commonRank: 80,
         examples: ['fork:yes repo:sourcegraph'],
     },
     {
-        type: FilterType.lang,
-        placeholder: parsePlaceholder('{language-name}'),
+        ...createQueryExampleFromString('{language-name}'),
+        field: FilterType.lang,
         description: 'Only include results from files in the specified programming language.',
         commonRank: 40,
         examples: ['lang:typescript encoding', '-lang:typescript encoding'],
     },
     {
-        type: FilterType.message,
-        placeholder: parsePlaceholder('"{any string}"'),
+        ...createQueryExampleFromString('"{any string}"'),
+        field: FilterType.message,
         description: `Only include results from diffs or commits which have commit messages containing the string.
 
 To use this filter, the search query must contain \`type:diff\` or \`type:commit\`.`,
         examples: ['type:commit message:"testing"', 'type:diff message:"testing"'],
     },
     {
-        type: FilterType.repo,
-        placeholder: parsePlaceholder('{regexp-pattern}'),
+        ...createQueryExampleFromString('{regexp-pattern}'),
+        field: FilterType.repo,
         description:
             'Only include results from repositories whose path matches the regexp-pattern. A repository’s path is a string such as *github.com/myteam/abc* or *code.example.com/xyz* that depends on your organization’s repository host. If the regexp ends in `@rev`, that revision is searched instead of the default branch (usually `master`). `repo:regexp-pattern@rev` is equivalent to `repo:regexp-pattern rev:rev`.',
         commonRank: 10,
@@ -171,74 +167,74 @@ To use this filter, the search query must contain \`type:diff\` or \`type:commit
         ],
     },
     {
-        type: FilterType.repogroup,
-        placeholder: parsePlaceholder('{group-name}'),
+        ...createQueryExampleFromString('{group-name}'),
+        field: FilterType.repogroup,
         description:
             'Only include results from the named group of repositories (defined by the server admin). Same as using a repo: keyword that matches all of the group’s repositories. Use repo: unless you know that the group exists.',
     },
     {
-        type: FilterType.repo,
-        placeholder: parsePlaceholder('contains.file({path})'),
+        ...createQueryExampleFromString('contains.file({path})'),
+        field: FilterType.repo,
         description: 'Search only inside repositories that contain a file path matching the regular expression.',
         examples: ['repo:contains.file(README)'],
         showSuggestions: false,
     },
     {
-        type: FilterType.repo,
-        placeholder: parsePlaceholder('contains.content({content})'),
+        ...createQueryExampleFromString('contains.content({content})'),
+        field: FilterType.repo,
         description: 'Search only inside repositories that contain file content matching the regular expression.',
         examples: ['repo:contains.content(TODO)'],
         showSuggestions: false,
     },
     {
-        type: FilterType.repo,
-        placeholder: parsePlaceholder('contains({file:path content:content})'),
+        ...createQueryExampleFromString('contains({file:path content:content})'),
+        field: FilterType.repo,
         description:
             'Search only inside repositories that contain a file matching the `file:` with `content:` filters.',
         examples: ['repo:contains(file:CHANGELOG content:fix)'],
         showSuggestions: false,
     },
     {
-        type: FilterType.repo,
-        placeholder: parsePlaceholder('contains.commit.after({date})'),
+        ...createQueryExampleFromString('contains.commit.after({date})'),
+        field: FilterType.repo,
         description:
             'Search only inside repositories that contain a a commit after some specified time. See [git date formats](https://github.com/git/git/blob/master/Documentation/date-formats.txt) for accepted formats. Use this to filter out stale repositories that don’t contain commits past the specified time frame. This parameter is experimental.',
         examples: ['repo:contains.commit.after(1 month ago)', 'repo:contains.commit.after(june 25 2017)'],
         showSuggestions: false,
     },
     {
-        type: FilterType.rev,
-        placeholder: parsePlaceholder('{revision}'),
+        ...createQueryExampleFromString('{revision}'),
+        field: FilterType.rev,
         commonRank: 20,
         description:
             'Search a revision instead of the default branch. `rev:` can only be used in conjunction with `repo:` and may not be used more than once. See our [revision syntax documentation](https://docs.sourcegraph.com/code_search/reference/queries#repository-revisions) to learn more.',
     },
     {
-        type: FilterType.select,
-        placeholder: parsePlaceholder('{result-types}'),
+        ...createQueryExampleFromString('{result-types}'),
+        field: FilterType.select,
         commonRank: 50,
         description:
             'Shows only query results for a given type. For example, `select:repo` displays only distinct reopsitory paths from search results. See [language definition](https://docs.sourcegraph.com/code_search/reference/language#select) for possible values.',
         examples: ['fmt.Errorf select:repo'],
     },
     {
-        type: FilterType.type,
-        placeholder: parsePlaceholder('{diff/commit/...}'),
+        ...createQueryExampleFromString('{diff/commit/...}'),
+        field: FilterType.type,
         commonRank: 1,
         description:
             'Specifies the type of search. By default, searches are executed on all code at a given point in time (a branch or a commit). Specify the `type:` if you want to search over changes to code or commit messages instead (diffs or commits).',
         examples: ['type:symbol path', 'type:diff func', 'type:commit test'],
     },
     {
-        type: FilterType.timeout,
-        placeholder: parsePlaceholder('{golang-duration-value}'),
+        ...createQueryExampleFromString('{golang-duration-value}'),
+        field: FilterType.timeout,
         description:
             'Customizes the timeout for searches. The value of the parameter is a string that can be parsed by the [Go time package’s `ParseDuration`](https://golang.org/pkg/time/#ParseDuration) (e.g. 10s, 100ms). By default, the timeout is set to 10 seconds, and the search will optimize for returning results as soon as possible. The timeout value cannot be set longer than 1 minute. When provided, the search is given the full timeout to complete.',
         examples: ['repo:^github.com/sourcegraph timeout:15s func count:10000'],
     },
     {
-        type: FilterType.visibility,
-        placeholder: parsePlaceholder('{any/private/public}'),
+        ...createQueryExampleFromString('{any/private/public}'),
+        field: FilterType.visibility,
         description:
             'Filter results to only public or private repositories. The default is to include both private and public repositories.',
         examples: ['type:repo visibility:public'],
@@ -256,25 +252,25 @@ const commonFilters = filterInfos
 
 const operatorInfo: OperatorInfo[] = [
     {
+        ...createQueryExampleFromString('{expr} AND {expr}'),
         operator: 'AND',
         alias: 'and',
-        placeholder: parsePlaceholder('{expr} AND {expr}'),
         description:
             'Returns results for files containing matches on the left and right side of the `and` (set intersection).',
         examples: ['conf.Get( and log15.Error(', 'conf.Get( AND log15.Error( AND after'],
     },
     {
+        ...createQueryExampleFromString('({expr} OR {expr})'),
         operator: 'OR',
         alias: 'or',
-        placeholder: parsePlaceholder('({expr} OR {expr})'),
         description:
             'Returns file content matching either on the left or right side, or both (set union). The number of results reports the number of matches of both strings.',
         examples: ['conf.Get( or log15.Error(', 'conf.Get( OR log15.Error( OR after'],
     },
     {
+        ...createQueryExampleFromString('NOT {expr}'),
         operator: 'NOT',
         alias: 'not',
-        placeholder: parsePlaceholder('NOT {expr}'),
         description:
             '`NOT` can be prepended to negate filters like `file`, `lang`, `repo`. Prepending `NOT` to search patterns excludes documents that contain the pattern. For readability, you may use `NOT` in conjunction with `AND` if you like: `panic AND NOT ever`.',
         examples: ['lang:go not file:main.go panic', 'panic NOT ever'],
@@ -286,7 +282,7 @@ const operatorInfo: OperatorInfo[] = [
  * filter information (name, description, ...)
  */
 function matches(searchTerms: RegExp[], info: FilterInfo): boolean {
-    return searchTerms.every(term => term.test(info.type) || term.test(info.description || ''))
+    return searchTerms.every(term => term.test(info.field) || term.test(info.description || ''))
 }
 
 /**
@@ -299,149 +295,20 @@ function parseSearchInput(searchInput: string): RegExp[] {
 }
 
 /**
- * Given a Placeholder object, this function returns the Monaco Selections
- * corresponding to each value in the Placeholder.
- */
-function selectionForPlaceholder(
-    placeholder: Placeholder,
-    offset: number = 0,
-    forSuggestions: boolean = false
-): Selection {
-    const token = placeholder.tokens.find(token => token.type === 'value')
-    if (!token) {
-        throw new Error('Search reference does not contain placeholder.')
-    }
-    const selectionStart = offset + token.start + 1
-    // For filters with suggestions we create an "empty" selection to position
-    // the cursor right after the colon.
-    const selectionEnd = forSuggestions ? selectionStart : offset + token.end
-    return new Selection(1, selectionStart, 1, selectionEnd)
-}
-
-/**
  * Whether or not to trigger the suggestion popover when adding this filter to
  * the query.
  */
 function shouldShowSuggestions(searchReference: FilterInfo): boolean {
-    return Boolean(searchReference.showSuggestions !== false && FILTERS[searchReference.type].discreteValues)
-}
-
-/**
- * This helper function will update the current query with the provided filter,
- * updating existing filters if necessary.
- * exported only for test purposes
- */
-export function updateQueryWithFilter(
-    currentQueryState: QueryState,
-    filterInfo: FilterInfo,
-    negate: boolean,
-    allFilters: typeof FILTERS
-): QueryState {
-    const { singular } = allFilters[filterInfo.type]
-    let { query } = currentQueryState
-    const showSuggestions = shouldShowSuggestions(filterInfo)
-    let selection: Selection | undefined
-    let revealRange: Range
-    let field: string = filterInfo.type
-
-    if (negate && isNegatableFilter(filterInfo.type)) {
-        field = '-' + field
-    }
-
-    const existingFilter = findFilter(query, filterInfo.type, FilterKind.Global)
-
-    if (existingFilter && singular) {
-        // Filter can only appear once
-        // Select or remove the existing filter value
-        if (showSuggestions) {
-            query = updateFilter(query, field, '')
-        }
-        const selectionStart = (existingFilter.value?.range.start || existingFilter.field.range.end) + 1
-        // For filters with suggestions, we create an "empty" selection to
-        // position the cursor after the colon
-        const selectionEnd = showSuggestions ? selectionStart : existingFilter.range.end + 1
-        selection = new Selection(1, selectionStart, 1, selectionEnd)
-        // A separate range is needed to make sure that the full filter, including
-        // the field name, is scrolled into view.
-        revealRange = new Range(1, existingFilter.range.start + 1, 1, selectionEnd)
-    } else {
-        // Filter can appear multiple times or doesn't exist yet. Always
-        // append.
-
-        // +1 because appendFilter inserts a whitespace character at the end of
-        // the query
-        const rangeStart = query.length + 1
-        query = appendFilter(query, field, showSuggestions ? '' : filterInfo.placeholder.text)
-        const offset = query.length - (showSuggestions ? 0 : filterInfo.placeholder.text.length)
-        selection = selectionForPlaceholder(
-            filterInfo.placeholder,
-            offset,
-            // If we need to trigger the suggestion popover we have to make
-            // sure the input cursor is positioned at the beginning of the
-            // selection (it usually is at the end)
-            showSuggestions
-        )
-        // A separate range is needed to make sure that the full filter, including
-        // the field name, is scrolled into view.
-        revealRange = new Range(1, rangeStart + 1, 1, selection.endColumn)
-    }
-
-    return {
-        changeSource: QueryChangeSource.searchReference,
-        query,
-        selection,
-        showSuggestions,
-        revealRange,
-    }
-}
-
-interface Placeholder {
-    tokens: { type: 'text' | 'value'; content: string; start: number; end: number }[]
-    text: string
-}
-
-export function parsePlaceholder(placeholder: string): Placeholder {
-    const valuePattern = /{([^}]+)}/g
-    let currentIndex = 0
-    const parsedPlaceholder: Placeholder = { tokens: [], text: '' }
-    let match
-    while ((match = valuePattern.exec(placeholder))) {
-        if (currentIndex !== match.index) {
-            parsedPlaceholder.tokens.push({
-                type: 'text',
-                content: placeholder.slice(currentIndex, match.index),
-                start: currentIndex,
-                end: match.index - 1,
-            })
-        }
-        parsedPlaceholder.tokens.push({
-            type: 'value',
-            content: match[1],
-            start: match.index,
-            end: match.index + match[0].length - 1,
-        })
-        currentIndex = match.index + match[0].length
-    }
-
-    if (currentIndex < placeholder.length) {
-        parsedPlaceholder.tokens.push({
-            type: 'text',
-            content: placeholder.slice(currentIndex),
-            start: currentIndex,
-            end: placeholder.length - 1,
-        })
-    }
-    parsedPlaceholder.text = parsedPlaceholder.tokens.map(token => token.content).join('')
-    return parsedPlaceholder
+    return Boolean(searchReference.showSuggestions !== false && FILTERS[searchReference.field].discreteValues)
 }
 
 function isFilterInfo(searchReference: SearchReferenceInfo): searchReference is FilterInfo {
-    return (searchReference as FilterInfo).type !== undefined
+    return (searchReference as FilterInfo).field !== undefined
 }
 
 const classNameTokenMap = {
     text: 'search-filter-keyword',
-    value: styles.placeholder,
+    placeholder: styles.placeholder,
 }
 
 interface SearchReferenceExampleProps {
@@ -497,7 +364,7 @@ const SearchReferenceEntry = <T extends SearchReferenceInfo>({
 
     let buttonTextPrefix: ReactElement | null = null
     if (isFilterInfo(searchReference)) {
-        buttonTextPrefix = <span className="search-filter-keyword">{searchReference.type}:</span>
+        buttonTextPrefix = <span className="search-filter-keyword">{searchReference.field}:</span>
     }
 
     return (
@@ -514,9 +381,9 @@ const SearchReferenceEntry = <T extends SearchReferenceInfo>({
                 >
                     <span className="text-monospace">
                         {buttonTextPrefix}
-                        {searchReference.placeholder.tokens.map(token => (
+                        {searchReference.tokens.map(token => (
                             <span key={token.start} className={classNameTokenMap[token.type]}>
-                                {token.content}
+                                {token.value}
                             </span>
                         ))}
                     </span>
@@ -548,9 +415,9 @@ const SearchReferenceEntry = <T extends SearchReferenceInfo>({
                             </span>
                         </p>
                     )}
-                    {isFilterInfo(searchReference) && isNegatableFilter(searchReference.type) && (
+                    {isFilterInfo(searchReference) && isNegatableFilter(searchReference.field) && (
                         <p>
-                            Negation: <span className="test-code search-filter-keyword">-{searchReference.type}:</span>
+                            Negation: <span className="test-code search-filter-keyword">-{searchReference.field}:</span>
                             {searchReference.alias && (
                                 <>
                                     {' '}
@@ -589,7 +456,7 @@ const FilterInfoList = ({ filters, onClick, onExampleClick }: FilterInfoListProp
     <ul className={styles.list}>
         {filters.map(filterInfo => (
             <SearchReferenceEntry
-                key={filterInfo.type + filterInfo.placeholder.text}
+                key={filterInfo.field + filterInfo.value}
                 searchReference={filterInfo}
                 onClick={onClick}
                 onExampleClick={onExampleClick}
@@ -628,7 +495,23 @@ const SearchReference = (props: SearchReferenceProps): ReactElement => {
 
     const updateQuery = useCallback(
         (searchReference: FilterInfo, negate: boolean) => {
-            onNavbarQueryChange(updateQueryWithFilter(navbarSearchQueryState, searchReference, negate, FILTERS))
+            const updatedQuery = updateQueryWithFilterAndExample(
+                navbarSearchQueryState.query,
+                searchReference.field,
+                searchReference,
+                {
+                    singular: Boolean(FILTERS[searchReference.field].singular),
+                    negate: negate && isNegatableFilter(searchReference.field),
+                    emptyValue: shouldShowSuggestions(searchReference),
+                }
+            )
+            onNavbarQueryChange({
+                changeSource: QueryChangeSource.searchReference,
+                query: updatedQuery.query,
+                selectionRange: updatedQuery.placeholderRange,
+                revealRange: updatedQuery.filterRange,
+                showSuggestions: shouldShowSuggestions(searchReference),
+            })
         },
         [onNavbarQueryChange, navbarSearchQueryState]
     )
@@ -677,7 +560,7 @@ const SearchReference = (props: SearchReferenceProps): ReactElement => {
                                 {operatorInfo.map(operatorInfo => (
                                     <SearchReferenceEntry
                                         searchReference={operatorInfo}
-                                        key={operatorInfo.operator + operatorInfo.placeholder.text}
+                                        key={operatorInfo.operator + operatorInfo.value}
                                         onClick={updateQueryWithOperator}
                                         onExampleClick={updateQueryWithExample}
                                     />