+++ enterprise/cmd/executor/internal/janitor/orphaned_vms.go
@@ -4,12 +4,12 @@ import (
 	"context"
 	"os/exec"
 	"sort"
-	"strings"
 	"time"
 
 	"github.com/hashicorp/go-multierror"
 	"github.com/inconshreveable/log15"
 
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/executor/internal/ignite"
 	"github.com/sourcegraph/sourcegraph/internal/goroutine"
 )
 
@@ -20,6 +20,7 @@ type orphanedVMJanitor struct {
 }
 
 var _ goroutine.Handler = &orphanedVMJanitor{}
+var _ goroutine.ErrorHandler = &orphanedVMJanitor{}
 
 // NewOrphanedVMJanitor returns a background routine that periodically removes all VMs
 // on the host that are not known by the worker running within this executor instance.
@@ -49,13 +50,13 @@ func newOrphanedVMJanitor(
 }
 
 func (j *orphanedVMJanitor) Handle(ctx context.Context) (err error) {
-	runningVMs, err := currentlyRunningVMs(ctx, j.prefix)
+	vmsByName, err := ignite.ActiveVMsByName(ctx, j.prefix, true)
 	if err != nil {
 		return err
 	}
 
-	for _, id := range findOrphanedVMs(runningVMs, j.names.Slice()) {
-		log15.Info("Removing orphaned vm", "id", id, "error", err)
+	for _, id := range findOrphanedVMs(vmsByName, j.names.Slice()) {
+		log15.Info("Removing orphaned VM", "id", id)
 
 		if removeErr := exec.CommandContext(ctx, "ignite", "rm", "-f", id).Run(); removeErr != nil {
 			err = multierror.Append(err, removeErr)
@@ -69,32 +70,7 @@ func (j *orphanedVMJanitor) Handle(ctx context.Context) (err error) {
 
 func (j *orphanedVMJanitor) HandleError(err error) {
 	j.metrics.numErrors.Inc()
-	log15.Error("Failed to clean up orphaned vms", "err", err)
-}
-
-// currentlyRunningVMs returns the set of VMs existant on the host as a map from VM names
-// to VM identifiers. VMs starting with a prefix distinct from the given prefix are ignored.
-func currentlyRunningVMs(ctx context.Context, prefix string) (map[string]string, error) {
-	cmd := exec.CommandContext(ctx, "ignite", "ps", "-a", "-t", "{{ .Name }}:{{ .UID }}")
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		return nil, err
-	}
-
-	return parseIgniteList(prefix, string(out)), nil
-}
-
-// parseIgniteList parses the output from the `ignite ps` invocation in currentlyRunningVMs.
-// VMs starting with a prefix distinct from the given prefix are ignored.
-func parseIgniteList(prefix, out string) map[string]string {
-	activeVMsMap := map[string]string{}
-	for _, line := range strings.Split(out, "\n") {
-		if parts := strings.Split(line, ":"); len(parts) == 2 && strings.HasPrefix(parts[0], prefix) {
-			activeVMsMap[parts[0]] = parts[1]
-		}
-	}
-
-	return activeVMsMap
+	log15.Error("Failed to remove up orphaned vms", "error", err)
 }
 
 // findOrphanedVMs returns the set of VM identifiers present in running VMs but