+++ enterprise/cmd/executor/main.go
@@ -2,21 +2,18 @@ package main
 
 import (
 	"context"
-	"fmt"
 	"log"
-	"net/http"
-	"time"
 
 	"github.com/inconshreveable/log15"
 	"github.com/opentracing/opentracing-go"
 	"github.com/prometheus/client_golang/prometheus"
 
+	"github.com/sourcegraph/sourcegraph/enterprise/cmd/executor/internal/ignite"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/executor/internal/janitor"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/executor/internal/worker"
 	"github.com/sourcegraph/sourcegraph/internal/debugserver"
 	"github.com/sourcegraph/sourcegraph/internal/env"
 	"github.com/sourcegraph/sourcegraph/internal/goroutine"
-	"github.com/sourcegraph/sourcegraph/internal/httpserver"
 	"github.com/sourcegraph/sourcegraph/internal/logging"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
 	"github.com/sourcegraph/sourcegraph/internal/trace"
@@ -61,15 +58,9 @@ func main() {
 			config.CleanupTaskInterval,
 			janitor.NewMetrics(observationContext),
 		))
-	}
-	if !config.DisableHealthServer {
-		routines = append(routines, httpserver.NewFromAddr(fmt.Sprintf(":%d", config.HealthServerPort), &http.Server{
-			ReadTimeout:  75 * time.Second,
-			WriteTimeout: 10 * time.Minute,
-			Handler:      httpserver.NewHandler(nil),
-		}))
-	}
 
+		mustRegisterVMCountMetric(observationContext, config.VMPrefix)
+	}
 	goroutine.MonitorBackgroundRoutines(context.Background(), routines...)
 }
 
@@ -83,4 +74,18 @@ func makeWorkerMetrics(queueName string) workerutil.WorkerMetrics {
 	return workerutil.NewMetrics(observationContext, "executor_processor", map[string]string{
 		"queue": queueName,
 	})
+}
+
+func mustRegisterVMCountMetric(observationContext *observation.Context, prefix string) {
+	observationContext.Registerer.MustRegister(prometheus.NewGaugeFunc(prometheus.GaugeOpts{
+		Name: "src_executor_vms_total",
+		Help: "Total number of running VMs.",
+	}, func() float64 {
+		runningVMsByName, err := ignite.ActiveVMsByName(context.Background(), prefix, false)
+		if err != nil {
+			log15.Error("Failed to determine number of running VMs", "error", err)
+		}
+
+		return float64(len(runningVMsByName))
+	}))
 }